changed user prof

+++ users/models.py
@@ -9,7 +9,7 @@ class Profile(models.Model):
     user = models.OneToOneField(User, on_delete=models.CASCADE)
     contact_no = models.TextField(null=False, blank=False)
     players = models.ManyToManyField(Player,null = True, blank = True)
-    total_score = models.IntegerField(default=0)
+    total_score = models.FloatField(default=0.0)
     prediction_score = models.FloatField(default=0.0)
     captain = models.ForeignKey(Player, on_delete=models.CASCADE, related_name = 'captain',null = True, blank = True)
     vice_captain = models.ForeignKey(Player,on_delete=models.CASCADE, related_name='vice_captain',null = True, blank = True)
